+++ examples/storage.py
@@ -1,11 +1,11 @@
 from scrapy import Spider, Request
 from scrapy.crawler import CrawlerProcess
-from scrapy_playwright.page import PageCoroutine
+from scrapy_playwright.page import PageMethod
 
 
 class StorageSpider(Spider):
     """
-    Set and get storage state
+    Set and get storage state. Also get the server's IP address.
     """
 
     name = "storage"
@@ -16,15 +16,21 @@ class StorageSpider(Spider):
             meta={
                 "playwright": True,
                 "playwright_include_page": True,
-                "playwright_page_coroutines": [
-                    PageCoroutine("evaluate_handle", "window.localStorage.setItem('foo', 'bar');"),
+                "playwright_page_methods": [
+                    PageMethod("evaluate_handle", "window.localStorage.setItem('foo', 'bar');"),
                 ],
             },
         )
 
     async def parse(self, response):
         page = response.meta["playwright_page"]
-        return {"url": response.url, "storage_state": await page.context.storage_state()}
+        storage_state = await page.context.storage_state()
+        await page.close()
+        return {
+            "url": response.url,
+            "storage_state": storage_state,
+            "ip_address": response.ip_address,
+        }
 
 
 if __name__ == "__main__":