+++ tests/test_browser_contexts.py
@@ -1,5 +1,5 @@
+import asyncio
 import platform
-import warnings
 
 import pytest
 from scrapy import Spider, Request
@@ -9,6 +9,54 @@ from tests.mockserver import StaticMockServer
 
 
 class MixinTestCaseMultipleContexts:
+    @pytest.mark.asyncio
+    async def test_contexts_max_pages_setting(self):
+        settings = {
+            "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
+            "PLAYWRIGHT_MAX_PAGES_PER_CONTEXT": 1234,
+        }
+        async with make_handler(settings) as handler:
+            assert handler.max_pages_per_context == 1234
+
+        settings = {"PLAYWRIGHT_BROWSER_TYPE": self.browser_type, "CONCURRENT_REQUESTS": 9876}
+        async with make_handler(settings) as handler:
+            assert handler.max_pages_per_context == 9876
+
+    @pytest.mark.asyncio
+    async def test_contexts_max_pages(self):
+        settings = {
+            "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
+            "PLAYWRIGHT_MAX_PAGES_PER_CONTEXT": 2,
+            "PLAYWRIGHT_CONTEXTS": {
+                "a": {"java_script_enabled": True},
+                "b": {"java_script_enabled": True},
+            },
+        }
+        async with make_handler(settings) as handler:
+            with StaticMockServer() as server:
+                requests = [
+                    handler._download_request(
+                        Request(
+                            server.urljoin(f"/index.html?a={i}"),
+                            meta={"playwright": True, "playwright_context": "a"},
+                        ),
+                        Spider("foo"),
+                    )
+                    for i in range(20)
+                ] + [
+                    handler._download_request(
+                        Request(
+                            server.urljoin(f"/index.html?b={i}"),
+                            meta={"playwright": True, "playwright_context": "b"},
+                        ),
+                        Spider("foo"),
+                    )
+                    for i in range(20)
+                ]
+                await asyncio.gather(*requests)
+
+            assert handler.stats.get_value("playwright/page_count/max_concurrent") == 4
+
     @pytest.mark.asyncio
     async def test_contexts_startup(self):
         settings = {
@@ -28,6 +76,9 @@ class MixinTestCaseMultipleContexts:
             },
         }
         async with make_handler(settings) as handler:
+            assert len(handler.contexts) == 1
+            assert len(handler.context_semaphores) == 1
+
             with StaticMockServer() as server:
                 meta = {
                     "playwright": True,
@@ -49,6 +100,8 @@ class MixinTestCaseMultipleContexts:
     @pytest.mark.asyncio
     async def test_contexts_dynamic(self):
         async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
+            assert len(handler.contexts) == 0
+            assert len(handler.context_semaphores) == 0
 
             with StaticMockServer() as server:
                 meta = {
@@ -70,6 +123,9 @@ class MixinTestCaseMultipleContexts:
                 req = Request(server.urljoin("/index.html"), meta=meta)
                 resp = await handler._download_request(req, Spider("foo"))
 
+            assert len(handler.contexts) == 1
+            assert len(handler.context_semaphores) == 1
+
             page = resp.meta["playwright_page"]
             storage_state = await page.context.storage_state()
             await page.close()
@@ -77,47 +133,6 @@ class MixinTestCaseMultipleContexts:
             assert cookie["name"] == "asdf"
             assert cookie["value"] == "qwerty"
             assert cookie["domain"] == "example.org"
-
-    @pytest.mark.asyncio
-    async def test_deprecated_setting(self):
-        settings = {
-            "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
-            "PLAYWRIGHT_CONTEXT_ARGS": {
-                "storage_state": {
-                    "cookies": [
-                        {
-                            "url": "https://example.org",
-                            "name": "asdf",
-                            "value": "qwerty",
-                        },
-                    ],
-                },
-            },
-        }
-        with warnings.catch_warnings(record=True) as warning_list:
-            async with make_handler(settings) as handler:
-                assert warning_list[0].category is DeprecationWarning
-                assert str(warning_list[0].message) == (
-                    "The PLAYWRIGHT_CONTEXT_ARGS setting is deprecated, please use"
-                    " PLAYWRIGHT_CONTEXTS instead. Keyword arguments defined in"
-                    " PLAYWRIGHT_CONTEXT_ARGS will be used when creating the 'default' context"
-                )
-
-                with StaticMockServer() as server:
-                    meta = {
-                        "playwright": True,
-                        "playwright_include_page": True,
-                    }
-                    req = Request(server.urljoin("/index.html"), meta=meta)
-                    resp = await handler._download_request(req, Spider("foo"))
-
-                page = resp.meta["playwright_page"]
-                storage_state = await page.context.storage_state()
-                await page.close()
-                cookie = storage_state["cookies"][0]
-                assert cookie["name"] == "asdf"
-                assert cookie["value"] == "qwerty"
-                assert cookie["domain"] == "example.org"
 
 
 class TestCaseMultipleContextsChromium(MixinTestCaseMultipleContexts):