+++ tests/test_playwright_requests.py
@@ -2,14 +2,22 @@ import json
 import logging
 import platform
 import subprocess
+import sys
+from ipaddress import ip_address
 from tempfile import NamedTemporaryFile
+from unittest.mock import MagicMock, patch
 
 import pytest
-from playwright.async_api import Dialog, Page as PlaywrightPage, TimeoutError
+from playwright.async_api import (
+    Dialog,
+    Page as PlaywrightPage,
+    TimeoutError as PlaywrightTimeoutError,
+)
 from scrapy import Spider, Request, FormRequest
+from scrapy.http.headers import Headers
 from scrapy.http.response.html import HtmlResponse
 
-from scrapy_playwright.page import PageCoroutine as PageCoro
+from scrapy_playwright.page import PageMethod
 
 from tests import make_handler
 from tests.mockserver import MockServer, StaticMockServer
@@ -20,6 +28,7 @@ def get_mimetype(file):
         ["file", "--mime-type", "--brief", file.name],
         stdout=subprocess.PIPE,
         universal_newlines=True,
+        check=False,
     ).stdout.strip()
 
 
@@ -58,7 +67,7 @@ class MixinTestCase:
         async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
             with MockServer() as server:
                 req = FormRequest(
-                    server.urljoin("/delay/2"), meta={"playwright": True}, formdata={"foo": "bar"}
+                    server.urljoin("/"), meta={"playwright": True}, formdata={"foo": "bar"}
                 )
                 resp = await handler._download_request(req, Spider("foo"))
 
@@ -69,14 +78,14 @@ class MixinTestCase:
             assert "Request body: foo=bar" in resp.text
 
     @pytest.mark.asyncio
-    async def test_page_coroutine_navigation(self):
+    async def test_page_method_navigation(self):
         async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
             with StaticMockServer() as server:
                 req = Request(
                     url=server.urljoin("/index.html"),
                     meta={
                         "playwright": True,
-                        "playwright_page_coroutines": [PageCoro("click", "a.lorem_ipsum")],
+                        "playwright_page_methods": [PageMethod("click", "a.lorem_ipsum")],
                     },
                 )
                 resp = await handler._download_request(req, Spider("foo"))
@@ -91,7 +100,7 @@ class MixinTestCase:
             assert text == "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
 
     @pytest.mark.asyncio
-    async def test_page_coroutine_infinite_scroll(self):
+    async def test_page_method_infinite_scroll(self):
         async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
             with StaticMockServer() as server:
                 req = Request(
@@ -99,12 +108,16 @@ class MixinTestCase:
                     headers={"User-Agent": "scrapy-playwright"},
                     meta={
                         "playwright": True,
-                        "playwright_page_coroutines": [
-                            PageCoro("wait_for_selector", selector="div.quote"),
-                            PageCoro("evaluate", "window.scrollBy(0, document.body.scrollHeight)"),
-                            PageCoro("wait_for_selector", selector="div.quote:nth-child(11)"),
-                            PageCoro("evaluate", "window.scrollBy(0, document.body.scrollHeight)"),
-                            PageCoro("wait_for_selector", selector="div.quote:nth-child(21)"),
+                        "playwright_page_methods": [
+                            PageMethod("wait_for_selector", selector="div.quote"),
+                            PageMethod(
+                                "evaluate", "window.scrollBy(0, document.body.scrollHeight)"
+                            ),
+                            PageMethod("wait_for_selector", selector="div.quote:nth-child(11)"),
+                            PageMethod(
+                                "evaluate", "window.scrollBy(0, document.body.scrollHeight)"
+                            ),
+                            PageMethod("wait_for_selector", selector="div.quote:nth-child(21)"),
                         ],
                     },
                 )
@@ -153,16 +166,47 @@ class MixinTestCase:
             assert handler.default_navigation_timeout == 0.5
 
     @pytest.mark.asyncio
-    async def test_timeout(self):
+    async def test_timeout_error(self, caplog):
         settings_dict = {
             "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
-            "PLAYWRIGHT_DEFAULT_NAVIGATION_TIMEOUT": 1000,
+            "PLAYWRIGHT_DEFAULT_NAVIGATION_TIMEOUT": 100,
         }
         async with make_handler(settings_dict) as handler:
             with MockServer() as server:
-                req = Request(server.urljoin("/delay/2"), meta={"playwright": True})
-                with pytest.raises(TimeoutError):
+                req = Request(server.urljoin("/delay/1"), meta={"playwright": True})
+                with pytest.raises(PlaywrightTimeoutError) as excinfo:
                     await handler._download_request(req, Spider("foo"))
+                assert (
+                    "scrapy-playwright",
+                    logging.WARNING,
+                    f"Closing page due to failed request: {req} ({type(excinfo.value)})",
+                ) in caplog.record_tuples
+
+    @pytest.mark.skipif(sys.version_info < (3, 8), reason="Fails on py37")
+    @patch("scrapy_playwright.handler.logger")
+    @pytest.mark.asyncio
+    async def test_route_continue_exception(self, logger):
+        """This test fails on Python 3.7 in the "logger.warning.assert_called_with" assertion,
+        but the "Captured log call" section shows the exact message that is expected :shrug:
+        """
+        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
+            req_handler = handler._make_request_handler("GET", Headers({}), body=None)
+            route = MagicMock()
+            playwright_request = MagicMock()
+            playwright_request.url = "https//example.org"
+
+            # safe error, only warn
+            exc = Exception("Target page, context or browser has been closed")
+            route.continue_.side_effect = exc
+            await req_handler(route, playwright_request)
+            logger.warning.assert_called_with(
+                f"{playwright_request}: failed processing Playwright request ({exc})"
+            )
+
+            # unknown error, re-raise
+            route.continue_.side_effect = ZeroDivisionError("asdf")
+            with pytest.raises(ZeroDivisionError):
+                await req_handler(route, playwright_request)
 
     @pytest.mark.asyncio
     async def test_context_kwargs(self):
@@ -178,16 +222,16 @@ class MixinTestCase:
                     url=server.urljoin("/scroll.html"),
                     meta={
                         "playwright": True,
-                        "playwright_page_coroutines": [
-                            PageCoro("wait_for_selector", selector="div.quote", timeout=1000),
+                        "playwright_page_methods": [
+                            PageMethod("wait_for_selector", selector="div.quote", timeout=1000),
                         ],
                     },
                 )
-                with pytest.raises(TimeoutError):
+                with pytest.raises(PlaywrightTimeoutError):
                     await handler._download_request(req, Spider("foo"))
 
     @pytest.mark.asyncio
-    async def test_page_coroutine_screenshot(self):
+    async def test_page_method_screenshot(self):
         async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
             with NamedTemporaryFile(mode="w+b") as png_file:
                 with StaticMockServer() as server:
@@ -195,19 +239,19 @@ class MixinTestCase:
                         url=server.urljoin("/index.html"),
                         meta={
                             "playwright": True,
-                            "playwright_page_coroutines": {
-                                "png": PageCoro("screenshot", path=png_file.name, type="png"),
+                            "playwright_page_methods": {
+                                "png": PageMethod("screenshot", path=png_file.name, type="png"),
                             },
                         },
                     )
                     await handler._download_request(req, Spider("foo"))
 
                 png_file.file.seek(0)
-                assert png_file.file.read() == req.meta["playwright_page_coroutines"]["png"].result
+                assert png_file.file.read() == req.meta["playwright_page_methods"]["png"].result
                 assert get_mimetype(png_file) == "image/png"
 
     @pytest.mark.asyncio
-    async def test_page_coroutine_pdf(self):
+    async def test_page_method_pdf(self):
         if self.browser_type != "chromium":
             pytest.skip("PDF generation is supported only in Chromium")
 
@@ -218,15 +262,15 @@ class MixinTestCase:
                         url=server.urljoin("/index.html"),
                         meta={
                             "playwright": True,
-                            "playwright_page_coroutines": {
-                                "pdf": PageCoro("pdf", path=pdf_file.name),
+                            "playwright_page_methods": {
+                                "pdf": PageMethod("pdf", path=pdf_file.name),
                             },
                         },
                     )
                     await handler._download_request(req, Spider("foo"))
 
                 pdf_file.file.seek(0)
-                assert pdf_file.file.read() == req.meta["playwright_page_coroutines"]["pdf"].result
+                assert pdf_file.file.read() == req.meta["playwright_page_methods"]["pdf"].result
                 assert get_mimetype(pdf_file) == "application/pdf"
 
     @pytest.mark.asyncio
@@ -262,10 +306,12 @@ class MixinTestCase:
     @pytest.mark.asyncio
     async def test_use_playwright_headers(self):
         """Ignore Scrapy headers"""
+        from scrapy_playwright.headers import use_playwright_headers
+
         settings_dict = {
             "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
             "PLAYWRIGHT_CONTEXTS": {"default": {"user_agent": self.browser_type}},
-            "PLAYWRIGHT_PROCESS_REQUEST_HEADERS": "scrapy_playwright.headers.use_playwright_headers",  # noqa: E501
+            "PLAYWRIGHT_PROCESS_REQUEST_HEADERS": use_playwright_headers,
         }
         async with make_handler(settings_dict) as handler:
             with MockServer() as server:
@@ -315,8 +361,8 @@ class MixinTestCase:
                     url=server.urljoin("/index.html"),
                     meta={
                         "playwright": True,
-                        "playwright_page_coroutines": [
-                            PageCoro("evaluate", "alert('foobar');"),
+                        "playwright_page_methods": [
+                            PageMethod("evaluate", "alert('foobar');"),
                         ],
                         "playwright_page_event_handlers": {
                             "dialog": spider.handle_dialog,
@@ -336,8 +382,8 @@ class MixinTestCase:
                     url=server.urljoin("/index.html"),
                     meta={
                         "playwright": True,
-                        "playwright_page_coroutines": [
-                            PageCoro("evaluate", "alert('foobar');"),
+                        "playwright_page_methods": [
+                            PageMethod("evaluate", "alert('foobar');"),
                         ],
                         "playwright_page_event_handlers": {
                             "dialog": "handle_dialog",
@@ -371,6 +417,52 @@ class MixinTestCase:
             " ignoring handler for event 'dialog'",
         ) in caplog.record_tuples
         assert getattr(spider, "dialog_message", None) is None
+
+    @pytest.mark.asyncio
+    async def test_response_attributes(self):
+        async with make_handler({"PLAYWRIGHT_BROWSER_TYPE": self.browser_type}) as handler:
+            with MockServer() as server:
+                spider = DialogSpider()
+                req = Request(
+                    url=server.urljoin("/index.html"),
+                    meta={"playwright": True},
+                )
+                response = await handler._download_request(req, spider)
+
+        assert response.ip_address == ip_address(server.address)
+
+    @pytest.mark.asyncio
+    async def test_abort_requests(self):
+        async def should_abort_request_async(request):
+            return request.resource_type == "image"
+
+        def should_abort_request_sync(request):
+            return request.resource_type == "image"
+
+        for predicate in (
+            lambda request: request.resource_type == "image",
+            should_abort_request_async,
+            should_abort_request_sync,
+        ):
+            settings_dict = {
+                "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
+                "PLAYWRIGHT_ABORT_REQUEST": predicate,
+            }
+            async with make_handler(settings_dict) as handler:
+                with StaticMockServer() as server:
+                    req = Request(
+                        url=server.urljoin("/gallery.html"),
+                        meta={"playwright": True},
+                    )
+                    await handler._download_request(req, Spider("foo"))
+
+                    req_prefix = "playwright/request_count"
+                    resp_prefix = "playwright/response_count"
+                    assert handler.stats.get_value(f"{req_prefix}/resource_type/document") == 1
+                    assert handler.stats.get_value(f"{req_prefix}/resource_type/image") == 3
+                    assert handler.stats.get_value(f"{resp_prefix}/resource_type/document") == 1
+                    assert handler.stats.get_value(f"{resp_prefix}/resource_type/image") is None
+                    assert handler.stats.get_value(f"{req_prefix}/aborted") == 3
 
 
 class TestCaseChromium(MixinTestCase):